Remove old ThemChoicePanel
