+++ res/css/views/settings/_ThemeChoicePanel.pcss
@@ -14,52 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-.mx_ThemeChoicePanel_themeSelectors {
-    color: $primary-content;
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-
-    > .mx_StyledRadioButton {
-        align-items: center;
-        padding: $font-16px;
-        box-sizing: border-box;
-        border-radius: 10px;
-        width: 180px;
-
-        background: $accent-200;
-        opacity: 0.4;
-
-        flex-shrink: 1;
-        flex-grow: 0;
-
-        margin-right: 15px;
-        margin-top: 10px;
-
-        font-weight: var(--cpd-font-weight-semibold);
-
-        > span {
-            justify-content: center;
-        }
-    }
-
-    > .mx_StyledRadioButton_enabled {
-        opacity: 1;
-
-        /* These colors need to be hardcoded because they don't change with the theme */
-        &.mx_ThemeSelector_light {
-            background-color: #f3f8fd;
-            color: #2e2f32;
-        }
-
-        &.mx_ThemeSelector_dark {
-            /* 5% lightened version of 181b21 */
-            background-color: #25282e;
-            color: #f3f8fd;
-        }
-    }
-}
-
 .mx_ThemeChoicePanel_ThemeSelectors {
     display: flex;
     flex-wrap: wrap;
