+++ src/components/views/settings/tabs/user/AppearanceUserSettingsTab.tsx
@@ -28,8 +28,7 @@ import { UIFeature } from "../../../../../settings/UIFeature";
 import { Layout } from "../../../../../settings/enums/Layout";
 import LayoutSwitcher from "../../LayoutSwitcher";
 import FontScalingPanel from "../../FontScalingPanel";
-import ThemeChoicePanel from "../../ThemeChoicePanel";
-import { ThemeChoicePanel as ThemeChoicePanel2 } from "../../ThemeChoicePanel2";
+import { ThemeChoicePanel } from "../../ThemeChoicePanel";
 import ImageSizePanel from "../../ImageSizePanel";
 import SettingsTab from "../SettingsTab";
 import { SettingsSection } from "../../shared/SettingsSection";
@@ -157,7 +156,6 @@ export default class AppearanceUserSettingsTab extends React.Component<IProps, I
             <SettingsTab data-testid="mx_AppearanceUserSettingsTab">
                 <SettingsSection>
                     <ThemeChoicePanel />
-                    <ThemeChoicePanel2 />
                     <LayoutSwitcher
                         userId={this.state.userId}
                         displayName={this.state.displayName}
